Added full Payara distro, removed jfrog from repositories, deploy once
Payara distro is now the full one instead of the embedded one. Embedded
one retained for custom testing.

Removed jfrog from the repositories since we only need jfrog to deploy
to, not to resolve artefacts from.

Deployment from travis to jfrog is now only done once instead of for
every server in the matrix.